+++ lua/sqlserver/adapters/sql_tools_service/constants.lua
@@ -1,0 +1,3 @@
+return {
+  client_name = "mssql_ls",
+}