+++ tests/connect_spec.lua
@@ -7,7 +7,7 @@ return {
   run_test_async = function()
     test_utils.ui_select_fake("master")
     sqlserver.connect()
-    local client = vim.lsp.get_clients({ name = "mssql_ls", bufnr = 0 })[1]
+    local client = test_utils.get_sql_client(0)
     local buf = vim.api.nvim_get_current_buf()
 
     -- The connect event is sent, then the intelliSenseReady event.