+++ tests/result_limits_spec.lua
@@ -28,7 +28,7 @@ OPTION (MAXRECURSION 150);
     utils.wait_for_schedule_async()
     sqlserver.execute_buffer()
 
-    local client = vim.lsp.get_clients({ name = "mssql_ls", bufnr = query_buffer })[1]
+    local client = test_utils.get_sql_client(query_buffer)
     local _, err = utils.wait_for_notification_async(query_buffer, client, "query/complete", 30000)
     assert(not err, err and err.message or "Query completion timed out")
     test_utils.defer_async(2000)