+++ tests/use_query_spec.lua
@@ -1,5 +1,6 @@
 local sqlserver = require("sqlserver")
 local utils = require("sqlserver.utils")
+local test_utils = require("tests.utils")
 local workspace_registry = require("sqlserver.core.workspace_registry")
 
 return {
@@ -9,7 +10,7 @@ return {
     vim.api.nvim_buf_set_lines(0, 0, -1, false, { query })
     utils.wait_for_schedule_async()
     sqlserver.execute_query()
-    local client = vim.lsp.get_clients({ name = "mssql_ls", bufnr = 0 })[1]
+    local client = test_utils.get_sql_client(0)
     local buf = vim.api.nvim_get_current_buf()
 
     local _, err = utils.wait_for_notification_async(buf, client, "query/complete", 30000)